Remove prop drilling
Remove prop drilling in Form component and replace with useContext
